pyproject.toml
[project]
name = "aita-v4"
version = "0.0.1"
description = "AGI"
authors = [
    { name = "Jaya Adithya Pavuluri", email = "jayaadithya2004@gmail.com" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
    "langgraph>=1.0.2",
    "langgraph-checkpoint-postgres>=2.0.0",
    "python-dotenv>=1.0.1",
    "langchain>=1.0.4",
    "pgvector>=0.3.5",
    "openai",
    "psycopg[binary]",
    "langchain-openai",
    "fastapi",
    "uvicorn"
]


[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["aita"]
[tool.setuptools.package-dir]
"aita" = "src/aita"


[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[dependency-groups]
dev = [
    "anyio>=4.7.0",
    "langgraph-cli[inmem]>=0.2.8",
    "mypy>=1.13.0",
    "pytest>=8.3.5",
    "ruff>=0.8.2",
]
